requirements.txt
opencv-python
numpy
mxnet-cu100
pandas